refactor: migrate clone to transaction
